requirements.txt
# A/B тестирование промптов: посты LR для Telegram
openai>=1.0.0
python-dotenv>=1.0.0
